[PR #295] merge SynchronousTermOccurrenceSaver with TermOccurrenceSaver interface and fix faulty throttle aspect test
