+++ src/main/java/cz/cvut/kbss/termit/service/document/TermOccurrenceSaver.java
@@ -1,16 +1,34 @@
 package cz.cvut.kbss.termit.service.document;
 
+import cz.cvut.kbss.termit.exception.TermItException;
 import cz.cvut.kbss.termit.model.Asset;
 import cz.cvut.kbss.termit.model.assignment.TermOccurrence;
+import cz.cvut.kbss.termit.persistence.dao.TermOccurrenceDao;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.concurrent.BlockingQueue;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
- * Saves occurrences of terms.
+ * Saves occurrences synchronously.
+ * <p>
+ * Existing occurrences are reused if they match.
  */
-public interface TermOccurrenceSaver {
+@Service
+public class TermOccurrenceSaver {
+
+    private static final Logger LOG = LoggerFactory.getLogger(TermOccurrenceSaver.class);
+
+    private final TermOccurrenceDao termOccurrenceDao;
+
+    public TermOccurrenceSaver(TermOccurrenceDao termOccurrenceDao) {
+        this.termOccurrenceDao = termOccurrenceDao;
+    }
 
     /**
      * Saves the specified occurrences of terms in the specified asset.
@@ -22,7 +40,24 @@ public interface TermOccurrenceSaver {
      * @param occurrences Occurrences to save
      * @param source      Asset in which the terms occur
      */
-    void saveOccurrences(List<TermOccurrence> occurrences, Asset<?> source);
+    @Transactional
+    public void saveOccurrences(List<TermOccurrence> occurrences, Asset<?> source) {
+        LOG.debug("Saving term occurrences for asset {}.", source);
+        removeAll(source);
+        LOG.trace("Persisting new occurrences in {}.", source);
+        occurrences.stream().filter(o -> !o.getTerm().equals(source.getUri())).forEach(termOccurrenceDao::persist);
+    }
+
+    public void saveOccurrence(TermOccurrence occurrence, Asset<?> source) {
+        if (occurrence.getTerm().equals(source.getUri())) {
+            return;
+        }
+        if(!termOccurrenceDao.exists(occurrence.getUri())) {
+            termOccurrenceDao.persist(occurrence);
+        } else {
+            LOG.debug("Occurrence already exists, skipping: {}", occurrence);
+        }
+    }
 
     /**
      * Continously saves occurrences from the queue while blocking current thread until
@@ -34,10 +69,31 @@ public interface TermOccurrenceSaver {
      * @param finished Whether all occurrences were added to the queue
      * @param toSave   the queue with occurrences to save
      */
-    void saveFromQueue(final Asset<?> source, final AtomicBoolean finished,
-                       final BlockingQueue<TermOccurrence> toSave);
-
-    void saveOccurrence(TermOccurrence occurrence, Asset<?> source);
+    @Transactional
+    public void saveFromQueue(final Asset<?> source, final AtomicBoolean finished,
+                              final BlockingQueue<TermOccurrence> toSave) {
+        LOG.debug("Saving term occurrences for asset {}.", source);
+        removeAll(source);
+        TermOccurrence occurrence;
+        long count = 0;
+        try {
+            while (!finished.get() || !toSave.isEmpty()) {
+                if (toSave.isEmpty()) {
+                    Thread.yield();
+                }
+                occurrence = toSave.poll(1, TimeUnit.SECONDS);
+                if (occurrence != null) {
+                    saveOccurrence(occurrence, source);
+                    count++;
+                }
+            }
+            LOG.debug("Saved {} term occurrences for assert {}.", count, source);
+        } catch (InterruptedException e) {
+            LOG.error("Thread interrupted while waiting for occurrences to save.");
+            Thread.currentThread().interrupt();
+            throw new TermItException(e);
+        }
+    }
 
     /**
      * Gets a list of existing term occurrences in the specified asset.
@@ -45,5 +101,12 @@ public interface TermOccurrenceSaver {
      * @param source Asset in which the terms occur
      * @return List of existing term occurrences
      */
-    List<TermOccurrence> getExistingOccurrences(Asset<?> source);
+    public List<TermOccurrence> getExistingOccurrences(Asset<?> source) {
+        return termOccurrenceDao.findAllTargeting(source);
+    }
+
+    private void removeAll(Asset<?> source) {
+        LOG.trace("Removing all existing occurrences in asset {}.", source);
+        termOccurrenceDao.removeAll(source);
+    }
 }
